+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - CI/CD action with Python testing, linting, and import checking
 - Database schemas for data provided by the NFL
+- Support for database migrations